[native_assets_cli] Prevent non src imports from src
