+++ pkgs/code_assets/pubspec.yaml
@@ -18,6 +18,9 @@ dependencies:
     path: ../hooks/
 
 dev_dependencies:
+  custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
   json_schema: ^5.2.0
+  repo_lint_rules:
+    path: ../repo_lint_rules/
   test: ^1.25.15
