+++ pkgs/hooks/pubspec.yaml
@@ -10,9 +10,12 @@ environment:
   sdk: '>=3.7.0 <4.0.0'
 
 dev_dependencies:
+  custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
   json_schema: ^5.2.0
   json_syntax_generator:
     path: ../json_syntax_generator/
   path: ^1.9.1
+  repo_lint_rules:
+    path: ../repo_lint_rules/
   test: ^1.25.15
