+++ pkgs/json_syntax_generator/analysis_options.yaml
@@ -7,6 +7,8 @@ analyzer:
     strict-casts: true
     strict-inference: true
     strict-raw-types: true
+  plugins:
+    - custom_lint
 
 linter:
   rules:
