+++ pkgs/json_syntax_generator/pubspec.yaml
@@ -18,6 +18,9 @@ dependencies:
   json_schema: ^5.2.0
 
 dev_dependencies:
+  custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
   path: ^1.9.1
+  repo_lint_rules:
+    path: ../repo_lint_rules/
   test: ^1.25.15
