+++ pkgs/native_assets_builder/lib/src/model/hook_result.dart
@@ -6,7 +6,9 @@ import 'package:collection/collection.dart';
 import 'package:native_assets_cli/native_assets_cli.dart';
 import 'package:native_assets_cli/native_assets_cli_builder.dart';
 
-import '../../native_assets_builder.dart';
+import '../build_runner/build_runner.dart';
+import 'build_result.dart';
+import 'link_result.dart';
 
 /// The result from a [NativeAssetsBuildRunner.build] or
 /// [NativeAssetsBuildRunner.link].
