+++ pkgs/native_assets_builder/pubspec.yaml
@@ -23,6 +23,10 @@ dependencies:
   yaml_edit: ^2.2.2
 
 dev_dependencies:
+  custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
   file_testing: ^3.0.2
+  repo_lint_rules:
+    path:
+      ../repo_lint_rules/
   test: ^1.25.15
