+++ pkgs/native_assets_cli/pubspec.yaml
@@ -25,7 +25,11 @@ dependencies:
   yaml: ^3.1.3 # Used for reading pubspec.yaml to obtain the package name
 
 dev_dependencies:
+  custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
   file_testing: ^3.0.2
   glob: any
+  repo_lint_rules:
+    path:
+      ../repo_lint_rules/
   test: ^1.25.15
