+++ pkgs/repo_lint_rules/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.1.0-wip
+
+- Initial version.
