+++ pubspec.yaml
@@ -58,6 +58,7 @@ workspace:
   - pkgs/native_assets_cli/example/link/app_with_asset_treeshaking
   - pkgs/native_assets_cli/example/link/package_with_assets
   - pkgs/native_toolchain_c
+  - pkgs/repo_lint_rules
 
 # Hook user-defines are specified in the pub workspace.
 hooks:
